backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
scikit-learn==1.4.0
numpy==1.26.2